+++ crates/openab-mcp/Cargo.toml
@@ -1,0 +1,50 @@
+[package]
+name = "openab-mcp"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+description = "OpenAB shared MCP runtime: outbound MCP client (OAuth, tool filters, circuit breaker) and the inbound OAB MCP Facade"
+
+[dependencies]
+tokio = { version = "1", features = ["full"] }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+reqwest = { version = "0.12", default-features = false, features = ["rustls-tls", "json"] }
+# rmcp 1.7 is on reqwest 0.13; its `StreamableHttpClient`/`AuthClient` impls are
+# for the 0.13 `Client`, so the OAuth transport must hold that type. The rest of
+# the crate stays on 0.12.
+reqwest013 = { package = "reqwest", version = "0.13", default-features = false, features = ["rustls", "json", "stream"] }
+anyhow = "1"
+async-trait = "0.1"
+tracing = "0.1"
+base64 = "0.22.1"
+sha2 = "0.11.0"
+getrandom = "0.4.2"
+urlencoding = "2.1.3"
+open = "5.3.5"
+url = "2.5.8"
+# Pinned to rmcp's oauth2 (5.0, default-features off) so `StoredCredentials`'
+# `OAuthTokenResponse` types unify; used to splice a rotated refresh_token back
+# into rmcp's wholesale-rebuilt credentials (OAuth 2.1 §10.4) in McpCredentialStore.
+oauth2 = { version = "5", default-features = false }
+# Multi-draft JSON Schema validator for MCP tool `inputSchema` (rows 18-21).
+jsonschema = { version = "0.46", default-features = false }
+rmcp = { version = "1.7", default-features = false, features = [
+    "client",
+    "transport-child-process",
+    "transport-streamable-http-client-reqwest",
+    "auth",
+    # OAB MCP Facade: inbound MCP server over loopback Streamable HTTP.
+    "server",
+    "transport-streamable-http-server",
+] }
+# Serves the facade's Streamable HTTP endpoint (rmcp's StreamableHttpService
+# is a tower Service; axum provides the loopback listener + router).
+axum = { version = "0.8", default-features = false, features = ["http1", "tokio"] }
+
+[target.'cfg(unix)'.dependencies]
+libc = "0.2"
+
+[dev-dependencies]
+tempfile = "3"
+temp-env = "0.3.6"
