+++ crates/openab-mcp/src/mcp/sampling.rs
